[refactor] DOM 조작 코드 winning-point.ts로 이동
