+++ src/pages/ingame/winning-point/winning-point.ts
@@ -77,18 +77,59 @@ export function getTotalPoints() {
   }));
 }
 
-// 실제 동작: 카드 제출 시점에 submitCard 호출
-// import { submitCard, getRoundResults, getTotalPoints } from "./winning-point";
+// HTML에 라운드별 우승자와 누적 승점 테이블을 렌더링하는 함수
+export function renderScoreTables() {
+  const roundResults = getRoundResults();
+  const totalPoints = getTotalPoints();
+
+  // 라운드별 결과 테이블 렌더링
+  const roundWinnerBody = document.getElementById("round-winner-body");
+  if (roundWinnerBody) {
+    roundWinnerBody.innerHTML = "";
+    roundResults.forEach(({ round, winners, point, draw }) => {
+      if (draw) {
+        const tr = document.createElement("tr");
+        tr.innerHTML = `
+        <td class="px-6 py-3 text-center">${round}</td>
+        <td class="px-6 py-3 text-center">무승부</td>
+        <td class="px-6 py-3 text-center">1</td>
+      `;
+        roundWinnerBody.appendChild(tr);
+      } else {
+        winners.forEach((winner) => {
+          const tr = document.createElement("tr");
+          tr.innerHTML = `
+          <td class="px-6 py-3 text-center">${round}</td>
+          <td class="px-6 py-3 text-center">${winner}</td>
+          <td class="px-6 py-3 text-center">${point}</td>
+        `;
+          roundWinnerBody.appendChild(tr);
+        });
+      }
+    });
+  }
+
+  // 누적 승점 테이블 렌더링
+  const totalScoreHeader = document.getElementById("total-score-header");
+  const totalScoreRow = document.getElementById("total-score-row");
+  if (totalScoreHeader && totalScoreRow) {
+    totalScoreHeader.innerHTML = totalPoints
+      .map((t) => `<th class="px-6 py-3">${t.nickName}</th>`)
+      .join("");
+    totalScoreRow.innerHTML = totalPoints
+      .map((t) => `<td class="px-6 py-3 text-center">${t.totalPoint}</td>`)
+      .join("");
+  }
+}
+
+// // 실제 동작: 카드 제출 시점에 submitCard 호출
+// import { submitCard, renderScoreTables } from "./winning-point";
 
 // // 예시: 서버에서 카드 제출 이벤트를 받았을 때
 // socket.on("cardSubmitted", (data) => {
 //   // data: { round, nickName, card }
 //   submitCard(data.round, data.nickName, data.card);
 
-//   // 라운드 결과/누적 승점 갱신
-//   const roundResults = getRoundResults();
-//   const totalPoints = getTotalPoints();
-
-//   // 결과를 UI에 반영 (테이블 갱신 등)
-//   updateScoreTables(roundResults, totalPoints);
+//   // 결과를 UI에 반영 (라운드 결과 및 누적 승점 테이블 갱신)
+//   renderScoreTables();
 // });
